refactor: change itemView to binding to hold view, and change bottomsheet to add view directory
