+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/modules/main/fragments/add/AddFragment.kt
@@ -29,7 +29,6 @@ import com.example.todo.R
 import com.example.todo.databinding.FragmentAddBinding
 import com.example.todo.modules.main.MainViewModel
 import com.example.todo.utils.bottomsheet.BaseBottomSheetFragment
-import com.example.todo.utils.bottomsheet.PhotoAccessBottomSheetFragment
 import com.example.todo.utils.listener.PhotoAccessListener
 import java.io.File
 import java.io.FileOutputStream
