+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/modules/main/fragments/home/HomeTaskViewHolder.kt
@@ -4,16 +4,15 @@ import android.app.AlertDialog
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.FrameLayout
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.todo.R
+import com.example.todo.databinding.ItemCardTaskBinding
 import com.example.todo.utils.listener.CardActionListener
 import com.example.todo.utils.models.Task
 
 class HomeTaskViewHolder(itemView: View, private val listener: CardActionListener) : RecyclerView.ViewHolder(itemView) {
 
+    private val binding: ItemCardTaskBinding = ItemCardTaskBinding.bind(itemView)
     private var dialog: AlertDialog? = null
     private var dialogView: View? = null
     private var task: Task? = null
@@ -22,11 +21,11 @@ class HomeTaskViewHolder(itemView: View, private val listener: CardActionListene
 
         this.task = task
 
-        dialogView = LayoutInflater.from(itemView.context)
-            .inflate(R.layout.item_dialog_task_check, itemView.parent as ViewGroup?, false)
+        dialogView = LayoutInflater.from(binding.root.context)
+            .inflate(R.layout.item_dialog_task_check, binding.root.parent as ViewGroup?, false)
 
-        dialog = AlertDialog.Builder(itemView.context)
-            .setCustomTitle(itemView.findViewById(R.id.titleDialog))
+        dialog = AlertDialog.Builder(binding.root.context)
+            .setCustomTitle(null)
             .setView(dialogView)
             .create()
 
@@ -37,23 +36,23 @@ class HomeTaskViewHolder(itemView: View, private val listener: CardActionListene
 
     private fun setupView() {
 
-        itemView.findViewById<TextView>(R.id.title).text = task?.title
-        itemView.findViewById<TextView>(R.id.content).text = task?.content
+        binding.title.text = task?.title
+        binding.content.text = task?.content
 
-        if(task?.image?.isEmpty() != true) {
+        if(task?.image != null) {
 
-            itemView.findViewById<ImageView>(R.id.image_icon).visibility = View.VISIBLE
-            itemView.findViewById<ImageView>(R.id.image_icon).setOnClickListener {
+            binding.imageIcon.visibility = View.VISIBLE
+
+            binding.imageIcon.setOnClickListener {
                 listener.onImageCLicked(task)
             }
 
         }
 
     }
-    private fun setupListener() {
 
-        itemView.findViewById<FrameLayout>(R.id.frame_layout).setOnClickListener { listener.onCheckClicked(task) }
+    private fun setupListener() =
+        binding.frameLayout.setOnClickListener { listener.onCheckClicked(task) }
 
-    }
 
 }
